refactor: player variable, see description
Previously, i was setting the player variable to null when the player left the interactables range, and setting the player variable to the player when the player entered the interactables range. This acted as a pseudo "in_range" flag, which was affecting both the 'player_is_in_line_of_sight' function and the 'is_interactable' function.

- the player variable now permanently stores the player.
- 'player_is_in_line_of_sight' now doesnt care if the interactable is in range or not, just if its in line of sight.
- 'is_interactable' now has an additional check for whether or not its in range of the player.